% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{binDaysByPhaseGroup}
\alias{binDaysByPhaseGroup}
\title{Assigns study phase (integer) with study day}
\usage{
binDaysByPhaseGroup(dat)
}
\arguments{
\item{dat}{vector of days}
}
\description{
Assigns study phase (integer) with study day
}
