% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps.R
\name{create_heatmap_iso}
\alias{create_heatmap_iso}
\title{Produces isolates heatmap}
\usage{
create_heatmap_iso(
  iso_profile,
  wgs_profile,
  mouse_ids = c("1683", "1688", "1692", "1699"),
  genome_hr = "C. innocuum",
  col_fun = col_fun,
  unfiltered = T
)
}
\arguments{
\item{iso_profile}{the AF profile of the genomic isolates}

\item{wgs_profile}{the AF profile of the WGS samples}

\item{genome_hr}{the name of the genome to plot, sould be present in iso_profile}

\item{col_fun}{function that assigns the AF color}

\item{unfiltered}{boolean, if intermediate AF should be filterd}
}
\value{
a ComplexHeamtap object
}
\description{
Produces isolates heatmap
}
