% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{group2color}
\alias{group2color}
\title{Assigns the color to a treatment group}
\usage{
group2color(dat)
}
\arguments{
\item{dat}{vector of treatment groups}
}
\description{
Assigns the color to a treatment group
}
