% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{theme_pmuench}
\alias{theme_pmuench}
\title{Custom ggplolt2 theme}
\usage{
theme_pmuench(base_size = 6, base_family = "")
}
\arguments{
\item{base_size}{size of text}

\item{base_family}{font type}
}
\description{
Custom ggplolt2 theme
}
