% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{translateGenomeIdToFullName}
\alias{translateGenomeIdToFullName}
\title{Translates the short gnome ID to human readable format}
\usage{
translateGenomeIdToFullName(dat)
}
\arguments{
\item{dat}{vector of short genome names}
}
\description{
Translates the short gnome ID to human readable format
}
