% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{translateIsolateIDtoBug}
\alias{translateIsolateIDtoBug}
\title{Assigns short genome ID to the}
\usage{
translateIsolateIDtoBug(dat)
}
\arguments{
\item{dat}{vector of days}
}
\description{
Assigns short genome ID to the
}
