% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{translateIsolateIDtoMouse}
\alias{translateIsolateIDtoMouse}
\title{Assigns a mouse ID to the isolate sample IDy}
\usage{
translateIsolateIDtoMouse(dat)
}
\arguments{
\item{dat}{vector of days}
}
\description{
Assigns a mouse ID to the isolate sample IDy
}
