% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{translateMouseIdToABC}
\alias{translateMouseIdToABC}
\title{Assigns one-letter code to mouse IDs for samples where WGS samples are present}
\usage{
translateMouseIdToABC(dat)
}
\arguments{
\item{dat}{vector of treatment groups}
}
\description{
Assigns one-letter code to mouse IDs for samples where WGS samples are present
}
