% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{translateMouseIdToTreatmentGroup}
\alias{translateMouseIdToTreatmentGroup}
\title{Assigns treatment group to mouse id}
\usage{
translateMouseIdToTreatmentGroup(dat)
}
\arguments{
\item{dat}{vector of mouse ids}
}
\description{
Assigns treatment group to mouse id
}
